Example of adding a mouse handler to craftEquipmentState (currently empty)
This handler should control whether to reset changes (left-click) or
open ufopedia for that item if it exist (right-click).

Right clicking for ufopedia was chosen due to consistency reasons (in known
existing actions it opens a new screen).

NOTE: probably not needed anymore since it looks Meridian already did it in his branch
(at least on the inventory).
NOTE: proposed branch name -> ufopeadiaFromInfoLists