Fix crash on extraSound loading (Option 1)
Do not throw an SDL error when we already know there is no audio-device available.

`Game.cpp` sets `Option::mute` if there is no audio-device we can connect to.
User cannot undo this during a play session.

Downside of this approach: we end up with a lot of `null` sounds.