Use same formula as Soldier Stat Screen for wounded soldiers.
The player will most likely compare it with that screen.
`getAccuracyModifier()` gives a different reading.
This discrepancy (between what is used for shooting and what the player sees on that screen) is the vanilla way.

I interpret the vanilla design decision this way:
Lets exaggerate the demise of accuracy on the stat screen, so the player would give some slack to a wounded soldier.