Implement functionality to move each visible item (vehicles excluded).
Added specialized buttons for this.

This functionality allows for a temporal overflow of items.
Any error from the overflow is stored in a queue and affects the OK button.
It will change appearance and functionality (show error messages).

To prevent players getting stuck a forced exit mode has been introduced.
* [ctrl]+[alt] click on OK button.