Fix ToggleButton changing appearance on unhandled mouse actions.
ToggleButton is used in 2 scenarios:
1) As bare button (no event attached).
2) As button with an event attached.

In the 2nd scenario the button could change appearance on an unhandled
event since it is triggered unconditionally on LMB or RMB (and only those).
This usually happens when using `onMouseClick` for defining handlers.
Since that one defaults to SDL_BUTTON_LEFT for the handler while
button appearance still changes on both LMB and RMB.

With this fix ToggleButton will change appearance on any mouseButton,
if attached as handler, while preserving the original bare scenario.

Including adaptions to existing code where it was implicitly expected
only LMB or RMB were allowed (or only LMB was considered).